requirements.txt
aiogram>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0
nest-asyncio>=1.5.5
httpx>=0.23.0
PyPDF2>=3.0.0
python-docx>=0.8.11
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.1
faiss-cpu>=1.7.4
numpy>=1.24.0
yookassa>=2.3.0
aiosqlite>=0.21.0
